Refactored generating and displaying graphs
